rename module cpu_usage to cpu
