+++ src/utils/__init__.py
@@ -1,5 +1,5 @@
 from utils.scraping import chrome_browser_setup
 from utils.scraping import get_carbon_intensity
-from utils.cpu_usage import get_cpu_info
-from utils.cpu_usage import get_cpu_usage
+from utils.cpu import get_cpu_info
+from utils.cpu import get_cpu_usage
 from utils.api_calls import get_request_co2signal
